Pretty good this time around - added a couple functions to work with WaveWatch III data - daydirs_ampspectra is useful for ambient noise studies - array_station_names added - good for extracting/defining arrays - started working on geofkarf - fixed hi-lat bug in fk functions (yay!) - fixed numerous minor bugs in fk functions - fkdbinfo takes several range options to get info about   a particular portion of the slowness space - geofkxchorzvolume now works (yay!) - plotgeofkmap no longer colors the land - lots of stability fixes to noise functions   - still need to figure out leap bug in daydirs_make - minor bugfixes or doc updates for the rest

+++ fk/plotfkmap.m
@@ -202,7 +202,7 @@ ny=numel(map.y);
 [x,y]=pol2cart(pi/180*map.x(ones(ny,1),:),map.y(:,ones(nx,1)));
 
 % plot polar grid
-pcolor(x,y,map.response);
+pcolor(x,y,double(map.response));
 
 % add title color etc
 hold off;
